refactor: Move generic handlers into generic directory
In preparation for introducing provider specific (AWS) patches.
