+++ cmd/main.go
@@ -27,15 +27,15 @@ import (
 	"github.com/d2iq-labs/capi-runtime-extensions/common/pkg/capi/clustertopology/handlers"
 	"github.com/d2iq-labs/capi-runtime-extensions/common/pkg/capi/clustertopology/handlers/mutation"
 	"github.com/d2iq-labs/capi-runtime-extensions/common/pkg/server"
-	"github.com/d2iq-labs/capi-runtime-extensions/pkg/handlers/auditpolicy"
-	"github.com/d2iq-labs/capi-runtime-extensions/pkg/handlers/clusterconfig"
-	"github.com/d2iq-labs/capi-runtime-extensions/pkg/handlers/cni/calico"
-	"github.com/d2iq-labs/capi-runtime-extensions/pkg/handlers/etcd"
-	"github.com/d2iq-labs/capi-runtime-extensions/pkg/handlers/extraapiservercertsans"
-	"github.com/d2iq-labs/capi-runtime-extensions/pkg/handlers/httpproxy"
-	"github.com/d2iq-labs/capi-runtime-extensions/pkg/handlers/kubernetesimagerepository"
-	"github.com/d2iq-labs/capi-runtime-extensions/pkg/handlers/nfd"
-	"github.com/d2iq-labs/capi-runtime-extensions/pkg/handlers/servicelbgc"
+	"github.com/d2iq-labs/capi-runtime-extensions/pkg/handlers/generic/clusterconfig"
+	"github.com/d2iq-labs/capi-runtime-extensions/pkg/handlers/generic/lifecycle/cni/calico"
+	"github.com/d2iq-labs/capi-runtime-extensions/pkg/handlers/generic/lifecycle/nfd"
+	"github.com/d2iq-labs/capi-runtime-extensions/pkg/handlers/generic/lifecycle/servicelbgc"
+	"github.com/d2iq-labs/capi-runtime-extensions/pkg/handlers/generic/mutation/auditpolicy"
+	"github.com/d2iq-labs/capi-runtime-extensions/pkg/handlers/generic/mutation/etcd"
+	"github.com/d2iq-labs/capi-runtime-extensions/pkg/handlers/generic/mutation/extraapiservercertsans"
+	"github.com/d2iq-labs/capi-runtime-extensions/pkg/handlers/generic/mutation/httpproxy"
+	"github.com/d2iq-labs/capi-runtime-extensions/pkg/handlers/generic/mutation/kubernetesimagerepository"
 )
 
 // Flags.
