+++ pkg/handlers/generic/lifecycle/cni/calico/handler.go
@@ -30,7 +30,7 @@ import (
 	"github.com/d2iq-labs/capi-runtime-extensions/common/pkg/k8s/client"
 	"github.com/d2iq-labs/capi-runtime-extensions/common/pkg/k8s/parser"
 	"github.com/d2iq-labs/capi-runtime-extensions/pkg/handlers"
-	"github.com/d2iq-labs/capi-runtime-extensions/pkg/handlers/cni"
+	"github.com/d2iq-labs/capi-runtime-extensions/pkg/handlers/generic/lifecycle/cni"
 )
 
 const (
