+++ pkg/handlers/generic/mutation/kubernetesimagerepository/variables.go
