+++ pkg/handlers/generic/mutation/kubernetesimagerepository/variables_test.go
